+++ PresentacionConsultaMovimientos/ConsultaMovimiento.cs
@@ -26,9 +26,9 @@ namespace PresentacionConsultaMovimientos
             textBoxNombreProducto.Text = "";
             textBoxNombreCortoProducto.Text = "";
             textBoxPrecioCosto.Text = "";
-            textboxStockActual.Text = "";
-            textBoxStockMinimo.Text = "";
-            textBoxPorcentajeGanancia.Text = "";
+            numericStockActual.Text = "";
+            numericStockMinimo.Text = "";
+            numericPorcentajeGanancia.Text = "";
         }
         public void actualizarTabla()
         {
@@ -51,7 +51,7 @@ namespace PresentacionConsultaMovimientos
         {
             // Funcion para agregar producto
             ConexionSQL_Negocio agregarProducto = new ConexionSQL_Negocio();
-            agregarProducto.agregarProducto(textboxCodigoProducto.Text, textBoxNombreProducto.Text, textBoxNombreCortoProducto.Text, textBoxPrecioCosto.Text, textboxStockActual.Text, textBoxStockMinimo.Text, textBoxPorcentajeGanancia.Text);
+            agregarProducto.agregarProducto(textboxCodigoProducto.Text, textBoxNombreProducto.Text, textBoxNombreCortoProducto.Text, textBoxPrecioCosto.Text, numericStockActual.Text, numericStockMinimo.Text, numericPorcentajeGanancia.Text);
             MessageBox.Show("Producto agregado correctamente");
             actualizarTabla();
             limpiarCampos();
@@ -64,10 +64,11 @@ namespace PresentacionConsultaMovimientos
             textboxCodigoProducto.Text = tablaProductos.SelectedCells[1].Value.ToString();
             textBoxNombreProducto.Text = tablaProductos.SelectedCells[2].Value.ToString();
             textBoxNombreCortoProducto.Text = tablaProductos.SelectedCells[3].Value.ToString();
-            textBoxPrecioCosto.Text = tablaProductos.SelectedCells[4].Value.ToString().Replace(",", ".");
-            textboxStockActual.Text = tablaProductos.SelectedCells[5].Value.ToString().Replace(",", ".");
-            textBoxStockMinimo.Text = tablaProductos.SelectedCells[6].Value.ToString().Replace(",", ".");
-            textBoxPorcentajeGanancia.Text = tablaProductos.SelectedCells[7].Value.ToString();
+            textBoxPrecioCosto.Text = tablaProductos.SelectedCells[4].Value.ToString();
+            numericStockActual.Text = tablaProductos.SelectedCells[5].Value.ToString();
+            numericStockMinimo.Text = tablaProductos.SelectedCells[6].Value.ToString();
+            numericPorcentajeGanancia.Text = tablaProductos.SelectedCells[7].Value.ToString();
+
 
 
 
@@ -82,22 +83,49 @@ namespace PresentacionConsultaMovimientos
         // Editar producto
         private void button1_Click(object sender, EventArgs e)
         {
-            ConexionSQL_Negocio editarProducto = new ConexionSQL_Negocio();
-            editarProducto.editarProducto(textboxID.Text, textboxCodigoProducto.Text, textBoxNombreProducto.Text, textBoxNombreCortoProducto.Text, textBoxPrecioCosto.Text, textboxStockActual.Text, textBoxStockMinimo.Text, textBoxPorcentajeGanancia.Text);
-            MessageBox.Show("Producto editado correctamente");
-            actualizarTabla();
-            limpiarCampos();
+
+            if (textboxID.Text == "")
+            {
+                MessageBox.Show("Seleccione un producto para editar");
+            }
+            else {
+                ConexionSQL_Negocio editarProducto = new ConexionSQL_Negocio();
+                string idProducto = textboxID.Text;
+                string codigoProducto = textboxCodigoProducto.Text;
+                string nombreProducto = textBoxNombreProducto.Text;
+                string nombreCortoProducto = textBoxNombreCortoProducto.Text;
+                string precioCosto = textBoxPrecioCosto.Text;
+                string stockActual = Convert.ToString(numericStockActual.Value);
+                string stockMinimo = Convert.ToString(numericStockMinimo.Value);
+                string porcentajeGanancia = Convert.ToString(numericPorcentajeGanancia.Value);
+
+                editarProducto.editarProducto(idProducto, codigoProducto, nombreProducto, nombreCortoProducto, precioCosto, stockActual, stockMinimo, porcentajeGanancia);
+
+                MessageBox.Show("Producto editado correctamente");
+                actualizarTabla();
+                limpiarCampos();
+            }
+
 
 
         }
 
         private void botonEliminarProducto_Click(object sender, EventArgs e)
         {
-            ConexionSQL_Negocio eliminarProducto = new ConexionSQL_Negocio();
-            eliminarProducto.eliminarProducto(textboxID.Text);        
-            MessageBox.Show("Producto eliminado correctamente");
-            actualizarTabla();
-            limpiarCampos();
+
+            if (textboxID.Text == "")
+            {
+                MessageBox.Show("Seleccione un producto para editar");
+            }
+            else {
+                ConexionSQL_Negocio eliminarProducto = new ConexionSQL_Negocio();
+                eliminarProducto.eliminarProducto(textboxID.Text);
+                MessageBox.Show("Producto eliminado correctamente");
+                actualizarTabla();
+                limpiarCampos();
+
+            }
+
 
         }
     }
